style: resolve some of Riley's comments
